+++ src/main/java/nextstep/main/java/nextstep/domain/reservation/ReservationCreateRequestDto.java
@@ -1,4 +1,4 @@
-package nextstep.main.java.nextstep.domain;
+package nextstep.main.java.nextstep.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
