+++ src/main/java/nextstep/main/java/nextstep/global/exception/exception/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package nextstep.main.java.nextstep.exception.exception;
+package nextstep.main.java.nextstep.global.exception.exception;
 
 public class ExceptionMessage {
     public static final String DUPLICATE_RESERVATION_MESSAGE = "날짜와 시간이 동일한 예약이 이미 존재합니다.";
