+++ src/main/java/nextstep/main/java/nextstep/repository/reservation/MemoryReservationRepository.java
@@ -1,6 +1,6 @@
-package nextstep.main.java.nextstep.repository;
+package nextstep.main.java.nextstep.repository.reservation;
 
-import nextstep.main.java.nextstep.domain.Reservation;
+import nextstep.main.java.nextstep.domain.reservation.Reservation;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
