+++ src/test/java/nextstep/main/java/nextstep/controller/ReservationControllerTest.java
@@ -2,9 +2,9 @@ package nextstep.main.java.nextstep.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.parsing.Parser;
-import nextstep.main.java.nextstep.domain.ReservationCreateRequestDto;
-import nextstep.main.java.nextstep.exception.exception.NoSuchReservationException;
-import nextstep.main.java.nextstep.service.ReservationService;
+import nextstep.main.java.nextstep.domain.reservation.ReservationCreateRequestDto;
+import nextstep.main.java.nextstep.global.exception.exception.NoSuchReservationException;
+import nextstep.main.java.nextstep.service.reservation.ReservationService;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
